pyTokens/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
influxdb-client==1.38.0
python-dateutil==2.8.2
python-dotenv==1.0.0
reactivex==4.0.4
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
